+++ docs/source/release_notes.rst
@@ -9,9 +9,14 @@ Future Release
         * Add ``MaxCount``, ``MedianCount``, ``MaxMinDelta``, ``NUniqueDays``, ``NMostCommonFrequency``,
             ``NUniqueDaysOfCalendarYear``, ``NUniqueDaysOfMonth``, ``NUniqueMonths``,
             ``NUniqueWeeks``, ``IsFirstWeekOfMonth`` (:pr:`2533`)
+        * Add ``HasNoDuplicates``, ``NthWeekOfMonth``, ``IsMonotonicallyDecreasing``, ``IsMonotonicallyIncreasing``,
+            ``IsUnique`` (:pr:`2537`)
     * Fixes
     * Changes
-        * Pinned pandas to 2.0.0 (:pr:`2533`)
+        * Restrict pandas to < 2.0.0 (:pr:`2533`)
+        * Upgrade minimum pandas to 1.5.0 (:pr:`2537`)
+        * Removed the ``Correlation`` and ``AutoCorrelation`` primitive as these could lead to data leakage (:pr:`2537`)
+        * Remove IntegerNullable support for ``Kurtosis`` primitive  (:pr:`2537`)
     * Documentation Changes
     * Testing Changes